GCC fixes
git-svn-id: http://svn.wildfiregames.com/public/ps/trunk@6191 3db68df2-c116-0410-a063-a993310a9797

+++ source/simulation/TerritoryManager.cpp
@@ -152,7 +152,7 @@ CTerritory* CTerritoryManager::GetTerritory(float x, float z)
 {
 	ssize_t ix, iz;
 	CTerrain::CalcFromPosition(x, z, ix, iz);
-	return GetTerritory(ix, iz);
+	return GetTerritory((int)ix, (int)iz);
 }
 
 // Calculate the boundary points of a given territory into the given vector
